+++ contentful/client.py
@@ -32,6 +32,8 @@ class Client(object):
     :param api_version: (optional) Target version of the Contentful API.
     :param default_locale: (optional) Default Locale for your Space,
         defaults to 'en-US'.
+    :param environment: (optional) Default Environment for client, defaults
+        to 'master'.
     :param https: (optional) Boolean determining wether to use https
         or http, defaults to True.
     :param authorization_as_header: (optional) Boolean determining wether
@@ -79,6 +81,7 @@ class Client(object):
             api_url='cdn.contentful.com',
             api_version=1,
             default_locale='en-US',
+            environment='master',
             https=True,
             authorization_as_header=True,
             raw_mode=False,
@@ -101,6 +104,7 @@ class Client(object):
         self.api_url = api_url
         self.api_version = api_version
         self.default_locale = default_locale
+        self.environment = environment
         self.https = https
         self.authorization_as_header = authorization_as_header
         self.raw_mode = raw_mode
@@ -156,7 +160,9 @@ class Client(object):
         """
 
         return self._get(
-            '/content_types/{0}'.format(content_type_id),
+            self.environment_url(
+                '/content_types/{0}'.format(content_type_id)
+            ),
             query
         )
 
@@ -178,7 +184,7 @@ class Client(object):
         """
 
         return self._get(
-            '/content_types',
+            self.environment_url('/content_types'),
             query
         )
 
@@ -204,7 +210,7 @@ class Client(object):
         try:
             query.update({'sys.id': entry_id})
             return self._get(
-                '/entries',
+                self.environment_url('/entries'),
                 query
             )[0]
         except IndexError:
@@ -240,7 +246,7 @@ class Client(object):
         self._normalize_select(query)
 
         return self._get(
-            '/entries',
+            self.environment_url('/entries'),
             query
         )
 
@@ -260,7 +266,9 @@ class Client(object):
         """
 
         return self._get(
-            '/assets/{0}'.format(asset_id),
+            self.environment_url(
+                '/assets/{0}'.format(asset_id)
+            ),
             query
         )
 
@@ -286,7 +294,30 @@ class Client(object):
         self._normalize_select(query)
 
         return self._get(
-            '/assets',
+            self.environment_url('/assets'),
+            query
+        )
+
+    def locales(self, query=None):
+        """Fetches all Locales from the Environment (up to the set limit, can be modified in `query`).
+
+        # TODO: fix url
+        API Reference: https://www.contentful.com/developers/docs/references/content-delivery-api/#/reference/assets/assets-collection/get-all-assets-of-a-space
+
+        :param query: (optional) Dict with API options.
+        :return: List of :class:`Locale <contentful.locale.Locale>` objects.
+        :rtype: List of contentful.locale.Locale
+
+        Usage:
+            >>> locales = client.locales()
+            [<Locale[English (United States)] code='en-US' default=True fallback_code=None optional=False>]
+        """
+
+        if query is None:
+            query = {}
+
+        return self._get(
+            self.environment_url('/locales'),
             query
         )
 
@@ -312,6 +343,15 @@ class Client(object):
             '/sync',
             query
         )
+
+    def environment_url(self, url):
+        """Formats the URL with the environment."""
+
+        return "/environments/{0}{1}".format(
+            self.environment,
+            url
+        )
+
 
     def _normalize_select(self, query):
         """